requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.6.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
numpy==2.1.2
Pillow==10.4.0
insightface==0.7.3
onnxruntime==1.19.2 # CPU
## onnxruntime-gpu==1.19.2 ## Gpu
psycopg[binary]==3.2.3
prometheus-fastapi-instrumentator==6.1.0
ddtrace==2.8.7
